fix: edit and branching
